requirements.txt
gpxpy
psycopg2-binary
python-dotenv
postgis
shapely
pandas
dash
dash-leaflet